requirements.txt
numpy==1.15.1
Pillow==10.1.0
torch==2.1.0
torchvision==0.16.0
